add date/time util function and refactor loading for reuse
