+++ src/components/loading.jsx
@@ -1,0 +1,14 @@
+const Loading = () => {
+  return (
+    <section>
+      <div className="block mt-6 pt-6">
+        <h5 className="has-text-info is-size-1 ml-3">nc news</h5>
+        <p className="is-size-2 mt-6">
+          One moment. <br></br>Loading...
+        </p>
+      </div>
+    </section>
+  );
+};
+
+export default Loading;
